+++ src/engine/db/dbformat.cc
@@ -102,12 +102,12 @@ const char* InternalFilterPolicy::Name() const {
   return user_policy_->Name();
 }
 
-void InternalFilterPolicy::CreateFilter(const Slice* keys, int n,
+void InternalFilterPolicy::CreateFilter(const Slice* keys, size_t n,
                                         std::string* dst) const {
   // We rely on the fact that the code in table.cc does not mind us
   // adjusting keys[].
   Slice* mkey = const_cast<Slice*>(keys);
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     mkey[i] = ExtractUserKey(keys[i]);
     // TODO(sanjay): Suppress dups?
   }
